Fix observer worker leak
