+++ src/storage/mod.rs
@@ -2,6 +2,7 @@ mod builder;
 mod config;
 mod core;
 mod observer;
+mod observer_worker;
 
 pub use self::{
     builder::Builder,
@@ -10,7 +11,10 @@ pub use self::{
 
 mod prelude {
     pub(crate) use {
-        super::{config::Config, core::Inner, observer::Observer},
+        super::{
+            config::Config, core::Inner, observer::Msg, observer::Observer,
+            observer_worker::ObserverWorker,
+        },
         crate::prelude::*,
     };
 }
